Update result script + some comments on polyline

+++ run_scenes.sh
@@ -1,9 +1,27 @@
 #!/bin/bash
 
+# arguments : 0 1 2 : run scene
+# t : run timings. f : make full video. c : copy to external drive
 
 ROOTDIR="/export/home/warhol/vroussel/code/git-clean/Radium-Engine"
 WORKDIR="/export/home/warhol/vroussel/code/git-clean/Radium-Engine/Bundle-GNU/RelWithDebInfo/bin"
 SCRIPTDIR="`date +%Y%m%d-%H%M`"
+COPYDIR="/mnt/STORM/wip/muscle-skinning/videos/source_files/"
+
+# returns true if string 1 contains string 2
+# than you https://stackoverflow.com/questions/2829613
+function contains()
+{
+    string="$1"
+    substring="$2"
+
+    if test "${string#*$substring}" != "$string"
+    then
+        return 0
+    else
+        return 1
+    fi
+}
 
 # run a scene exporting subdivided meshes and anatomy
 function run_scene_exportmesh()
@@ -118,40 +136,63 @@ function run_scene()
 }
 
 
+args="'$*'"
 echo "Starting script"
 cd $WORKDIR
 
 mkdir -p $SCRIPTDIR
 
-echo "Running scene 0 (parameters presentation)"
-run_scene "00" "/home/vroussel/Downloads/GEOM/arm_nomove" "scene00"
-run_scene "01" "/home/vroussel/Downloads/GEOM/arm_nomove" "scene01"
+if $(contains "$args" "0")
+then
+    echo "Running scene 0 (parameters presentation)"
+    run_scene "00" "/home/vroussel/Downloads/GEOM/arm_nomove" "scene00"
+    run_scene "01" "/home/vroussel/Downloads/GEOM/arm_nomove" "scene01"
+fi
 
-echo "Running scene 1 (arm shake)"
-run_scene 1 "/home/vroussel/Downloads/GEOM/arm_rise_shake" "scenefile"
+if $(contains "$args" "1")
+then
+    echo "Running scene 1 (arm shake)"
+    run_scene 1 "/home/vroussel/Downloads/GEOM/arm_rise_shake" "scenefile"
+fi
 
+if $(contains "$args" "2")
+then
+    echo "Running scene 2 (biceps curl)"
+    run_scene 2 "/home/vroussel/Downloads/GEOM/arm_biceps_curl" "scenefile"
+fi
 
-echo "Running scene 2 (biceps curl)"
-run_scene 2 "/home/vroussel/Downloads/GEOM/arm_biceps_curl" "scenefile"
-
-rm -f list.txt
-touch list.txt
-echo "Making full video"
-find $SCRIPTDIR | grep mp4 | sort | while read file
-do
-    echo "file ' $file'" >> list.txt
-done
-ffmpeg -f concat -safe 0 -i list.txt -c copy fullvideo.mp4
-mv fullvideo.mp4 $SCRIPTDIR
-
-echo "Running timings on scene 2"
-run_scene_timings "/home/vroussel/Downloads/GEOM/arm_biceps_curl" "$SCRIPTDIR/timings" "scenefile"
-# do not cat the first frame because the setup task screw up the stats
-find "$SCRIPTDIR/timings/on"  | grep radiumtimings | grep -v 000000 | xargs cat | $ROOTDIR/parse_results.py >$SCRIPTDIR/timings_on.txt
-find "$SCRIPTDIR/timings/off" | grep radiumtimings | grep -v 000000 | xargs cat  |$ROOTDIR/parse_results.py >$SCRIPTDIR/timings_off.txt
+if $(contains "$args" "3")
+then
+    echo "Running scene 3 (Pectorals)"
+    run_scene 3 "/home/vroussel/Downloads/GEOM/arm_pecs_biceps" "scenefile"
+fi
 
 
-echo "Copying"
-cp -a $SCRIPTDIR /mnt/STORM/wip/muscle-skinning/videos/source_files/
+if $(contains "$args" "f")
+then
+    rm -f list.txt
+    touch list.txt
+    echo "Making full video"
+    find $SCRIPTDIR | grep mp4 | sort | while read file
+    do
+        echo "file ' $file'" >> list.txt
+    done
+    ffmpeg -f concat -safe 0 -i list.txt -c copy fullvideo.mp4
+    mv fullvideo.mp4 $SCRIPTDIR
+fi
 
 
+if $(contains "$args" "t")
+then
+    echo "Running timings on scene 2"
+    run_scene_timings "/home/vroussel/Downloads/GEOM/arm_biceps_curl" "$SCRIPTDIR/timings" "scenefile"
+    # do not cat the first frame because the setup task screw up the stats
+    find "$SCRIPTDIR/timings/on"  | grep radiumtimings | grep -v 000000 | xargs cat | $ROOTDIR/parse_results.py >$SCRIPTDIR/timings_on.txt
+    find "$SCRIPTDIR/timings/off" | grep radiumtimings | grep -v 000000 | xargs cat  |$ROOTDIR/parse_results.py >$SCRIPTDIR/timings_off.txt
+fi
+
+if $(contains "$args" "c")
+then
+    echo "Copying"
+    cp -a $SCRIPTDIR $COPYDIR
+fi
